+++ app/core/prom_lib.py
@@ -1,0 +1,69 @@
+import os
+import settings
+
+class prometheus():
+    def __init__(self):
+        pass
+    
+    def emit(self,input_dict):
+        """
+        DESC: 
+        INPUT: 
+        OUTPUT: None
+        NOTE: None
+        """
+        pass
+        
+    def _used_memory(self):
+        """
+        DESC: 
+        INPUT: 
+        OUTPUT: None
+        NOTE: None
+        """
+        pass
+    
+    def _free_memory(self):
+        """
+        DESC: 
+        INPUT: 
+        OUTPUT: None
+        NOTE: None
+        """
+        pass
+    
+    def _used_swap(self):
+        """
+        DESC: 
+        INPUT: 
+        OUTPUT: None
+        NOTE: None
+        """
+        pass
+        
+    def _free_swap(self):
+        """
+        DESC: 
+        INPUT: 
+        OUTPUT: None
+        NOTE: None
+        """
+        pass
+        
+    def _upload(self):
+        """
+        DESC: 
+        INPUT: 
+        OUTPUT: None
+        NOTE: None
+        """
+        pass
+    
+    def _download(self):
+        """
+        DESC: 
+        INPUT: 
+        OUTPUT: None
+        NOTE: None
+        """
+        pass
